abstracted time tools to a new file
